refactor(components): refactor togglebutton
fix #283
