+++ packages/components/src/base/ToggleButtonGroup/index.tsx
@@ -1,0 +1,1 @@
+export { ToggleButtonGroup } from './ToggleButtonGroup';
